requirements.txt
annotated-types==0.7.0
anyio==4.5.2
cffi==1.17.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.2
dnspython==2.6.1
ecdsa==0.19.1
exceptiongroup==1.2.2
fastapi==0.115.12
fuzzywuzzy==0.18.0
h11==0.14.0
idna==3.10
Levenshtein==0.25.1
numpy==1.24.4
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
packaging==24.2
pillow==10.4.0
pyasn1==0.4.8
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pymongo==4.10.1
pytesseract==0.3.13
python-dotenv==1.0.1
python-jose==3.4.0
python-Levenshtein==0.25.1
python-multipart==0.0.20
rapidfuzz==3.9.7
rsa==4.9
six==1.17.0
sniffio==1.3.1
starlette==0.44.0
typing_extensions==4.13.1
uvicorn==0.33.0
